Fixed again warnings when using the Intel Composer XE 2013 (sp1.1.106) with -Wall -w2 -Wcheck
Reported-by: Markus Bina <bina(at)iue.tuwien.ac.at>

+++ viennacl/tools/timer.hpp
@@ -83,7 +83,7 @@ namespace viennacl{
       {
         struct timeval tval;
         gettimeofday(&tval, NULL);
-        ts = tval.tv_sec * 1000000 + tval.tv_usec;
+        ts = static_cast<double>(tval.tv_sec * 1000000 + tval.tv_usec);
       }
 
       double get() const
